Refactor Google Workspace Rust contracts into modules
