+++ crates/source-runtime-next/src/google_workspace.rs
@@ -8,17 +8,17 @@
 //! This fails closed for fallback-only records that Go Read can emit but Go
 //! Discover cannot turn into a canonical source URN.
 
-use std::{
-    collections::{BTreeMap, BTreeSet},
-    error::Error,
-    fmt,
-    net::IpAddr,
-    str::FromStr,
-};
+use std::{collections::BTreeMap, error::Error, fmt, str::FromStr};
 
 use reqwest::Url;
-use serde_json::{Map, Value};
-use time::{OffsetDateTime, UtcOffset, format_description::well_known::Rfc3339};
+use serde_json::Value;
+
+mod materialization;
+mod normalization;
+#[cfg(test)]
+mod tests;
+
+use normalization::*;
 
 const SOURCE_ID: &str = "google_workspace";
 const DEFAULT_CUSTOMER_ID: &str = "my_customer";
@@ -152,33 +152,6 @@ pub struct GoogleWorkspaceRecord {
     pub payload: Value,
 }
 
-/// One Google Workspace event materialized at page observation time.
-///
-/// Event fields match the Go source after the stricter Discover identity gate
-/// succeeds. This is a deliberate fail-closed tightening, not unconditional
-/// Read parity for provider records that omit their canonical source identity.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct GoogleWorkspaceEvent {
-    /// Stable source event identity.
-    pub event_id: String,
-    /// Tenant domain from source settings.
-    pub tenant_id: String,
-    /// Source identifier.
-    pub source_id: String,
-    /// Emitted provider kind.
-    pub provider_kind: String,
-    /// Public event schema reference.
-    pub schema_ref: String,
-    /// Go-selected occurrence time, normalized to UTC RFC 3339.
-    pub occurred_at: String,
-    /// Discovery URN produced from the Go Discover identity contract.
-    pub discovery_urn: String,
-    /// Source attributes derived from the provider object.
-    pub attributes: BTreeMap<String, String>,
-    /// Provider object with the Go source's context overlay.
-    pub payload: Value,
-}
-
 /// A bounded Google Workspace page and its opaque provider cursor.
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct GoogleWorkspacePage {
@@ -186,60 +159,6 @@ pub struct GoogleWorkspacePage {
     pub records: Vec<GoogleWorkspaceRecord>,
     /// `nextPageToken` for the next provider page.
     pub next_cursor: Option<String>,
-}
-
-impl GoogleWorkspacePage {
-    /// Materialize event, checkpoint, and watermark semantics.
-    ///
-    /// `observed_at` supplies the clock value used by the Go source for groups,
-    /// memberships, role assignments, and invalid or absent provider times.
-    /// Records that have only a Go Read fallback identity fail closed because
-    /// the same record would be rejected by the Go Discover URN path.
-    pub fn materialize(
-        &self,
-        observed_at: &str,
-    ) -> Result<GoogleWorkspaceEventPage, GoogleWorkspaceError> {
-        let observed_at = parse_observed_at(observed_at)?;
-        if self.records.is_empty() {
-            return Ok(GoogleWorkspaceEventPage {
-                events: Vec::new(),
-                next_cursor: None,
-                checkpoint_cursor: None,
-                watermark: None,
-            });
-        }
-        let events = self
-            .records
-            .iter()
-            .map(|record| materialize_event(record, observed_at))
-            .collect::<Result<Vec<_>, _>>()?;
-        let checkpoint_cursor = self.next_cursor.clone().or_else(|| {
-            events
-                .last()
-                .map(|event| event.event_id.trim().to_owned())
-                .filter(|value| !value.is_empty())
-        });
-        let watermark = events.last().map(|event| event.occurred_at.clone());
-        Ok(GoogleWorkspaceEventPage {
-            events,
-            next_cursor: self.next_cursor.clone(),
-            checkpoint_cursor,
-            watermark,
-        })
-    }
-}
-
-/// One materialized source page without durable checkpoint ownership.
-#[derive(Clone, Debug, Eq, PartialEq)]
-pub struct GoogleWorkspaceEventPage {
-    /// Go-compatible events in provider order.
-    pub events: Vec<GoogleWorkspaceEvent>,
-    /// Opaque provider cursor exposed for the next read.
-    pub next_cursor: Option<String>,
-    /// Proposed checkpoint cursor: provider cursor, then final event identity.
-    pub checkpoint_cursor: Option<String>,
-    /// Final event occurrence time used by the Go source as its watermark.
-    pub watermark: Option<String>,
 }
 
 /// Result of decoding one Google Workspace response.
@@ -560,1016 +479,3 @@ impl fmt::Display for GoogleWorkspaceError {
 }
 
 impl Error for GoogleWorkspaceError {}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum RequestStage {
-    Direct,
-    ResolveRoleUser,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct RoleAssignmentState {
-    records: Vec<Value>,
-    next_cursor: Option<String>,
-    lookups: Vec<String>,
-    lookup_index: usize,
-    users: BTreeMap<String, ResolvedUser>,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq)]
-struct ResolvedUser {
-    email: String,
-    name: String,
-}
-
-struct DecodedPage {
-    records: Vec<Value>,
-    next_cursor: Option<String>,
-}
-
-fn validate_origin(raw: &str) -> Result<Url, GoogleWorkspaceError> {
-    let mut url = Url::parse(raw.trim()).map_err(|_| GoogleWorkspaceError::InvalidBaseUrl)?;
-    let host = url.host_str().ok_or(GoogleWorkspaceError::InvalidBaseUrl)?;
-    let loopback =
-        host == "localhost" || IpAddr::from_str(host).is_ok_and(|address| address.is_loopback());
-    if url.scheme() != "https" && !(url.scheme() == "http" && loopback) {
-        return Err(GoogleWorkspaceError::InvalidBaseUrl);
-    }
-    if !url.username().is_empty()
-        || url.password().is_some()
-        || url.query().is_some()
-        || url.fragment().is_some()
-        || !matches!(url.path(), "" | "/")
-    {
-        return Err(GoogleWorkspaceError::InvalidBaseUrl);
-    }
-    if IpAddr::from_str(host).is_ok_and(|address| unsafe_ip_literal(address, loopback)) {
-        return Err(GoogleWorkspaceError::InvalidBaseUrl);
-    }
-    if url.port().is_some_and(|port| port != 443) && !loopback {
-        return Err(GoogleWorkspaceError::InvalidBaseUrl);
-    }
-    url.set_path("/");
-    Ok(url)
-}
-
-fn unsafe_ip_literal(address: IpAddr, loopback: bool) -> bool {
-    if loopback {
-        return false;
-    }
-    match address {
-        IpAddr::V4(address) => {
-            address.is_private()
-                || address.is_link_local()
-                || address.is_broadcast()
-                || address.is_documentation()
-                || address.is_unspecified()
-                || address.is_multicast()
-        }
-        IpAddr::V6(address) => {
-            address.is_unique_local()
-                || address.is_unicast_link_local()
-                || address.is_unspecified()
-                || address.is_multicast()
-        }
-    }
-}
-
-fn required_value(
-    value: &str,
-    error: GoogleWorkspaceError,
-) -> Result<String, GoogleWorkspaceError> {
-    let value = value.trim();
-    if value.is_empty() {
-        return Err(error);
-    }
-    Ok(value.to_owned())
-}
-
-fn optional_value(value: Option<String>, default: &str) -> String {
-    nonblank(value).unwrap_or_else(|| default.to_owned())
-}
-
-fn nonblank(value: Option<String>) -> Option<String> {
-    value
-        .map(|value| value.trim().to_owned())
-        .filter(|value| !value.is_empty())
-}
-
-fn decode_page(body: &[u8], field: &str) -> Result<DecodedPage, GoogleWorkspaceError> {
-    let root = decode_object(body)?;
-    let records = match root.get(field) {
-        None | Some(Value::Null) => Vec::new(),
-        Some(Value::Array(records)) => records.clone(),
-        Some(_) => return Err(GoogleWorkspaceError::InvalidResponse),
-    };
-    let next_cursor = match root.get("nextPageToken") {
-        None | Some(Value::Null) => None,
-        Some(Value::String(value)) => {
-            let value = value.trim();
-            (!value.is_empty()).then(|| value.to_owned())
-        }
-        Some(_) => return Err(GoogleWorkspaceError::InvalidResponse),
-    };
-    Ok(DecodedPage {
-        records,
-        next_cursor,
-    })
-}
-
-fn decode_object(body: &[u8]) -> Result<Map<String, Value>, GoogleWorkspaceError> {
-    serde_json::from_slice::<Value>(body)
-        .map_err(|_| GoogleWorkspaceError::InvalidResponse)?
-        .as_object()
-        .cloned()
-        .ok_or(GoogleWorkspaceError::InvalidResponse)
-}
-
-fn role_user_lookups(records: &[Value]) -> Result<Vec<String>, GoogleWorkspaceError> {
-    let mut seen = BTreeSet::new();
-    let mut lookups = Vec::new();
-    for record in records {
-        let object = record
-            .as_object()
-            .ok_or(GoogleWorkspaceError::InvalidRecord)?;
-        if scalar(object.get("assigneeType"))
-            .is_some_and(|value| value.eq_ignore_ascii_case("user"))
-            && let Some(user_key) = scalar(object.get("assignedTo"))
-            && !user_key.is_empty()
-            && seen.insert(user_key.clone())
-        {
-            lookups.push(user_key);
-        }
-    }
-    Ok(lookups)
-}
-
-fn normalize_user(
-    object: &Map<String, Value>,
-    fields: &mut BTreeMap<String, String>,
-) -> Result<String, GoogleWorkspaceError> {
-    let provider_id = first_values(object, &["id", "primaryEmail"])
-        .ok_or(GoogleWorkspaceError::MissingRecordIdentity)?;
-    let email = scalar(object.get("primaryEmail")).unwrap_or_default();
-    insert(
-        fields,
-        "user_id",
-        scalar(object.get("id")).unwrap_or_default(),
-    );
-    insert(fields, "primary_email", email.clone());
-    insert(fields, "email", email.clone());
-    insert(fields, "login", email);
-    insert(
-        fields,
-        "display_name",
-        nested_scalar(object, &["name", "fullName"]).unwrap_or_default(),
-    );
-    for (source, target) in [
-        ("creationTime", "created_at"),
-        ("lastLoginTime", "last_login_at"),
-        ("isAdmin", "is_admin"),
-        ("isDelegatedAdmin", "is_delegated_admin"),
-        ("isEnrolledIn2Sv", "mfa_enrolled"),
-        ("isEnforcedIn2Sv", "mfa_enforced"),
-        ("suspended", "suspended"),
-        ("archived", "archived"),
-        ("orgUnitPath", "org_unit_path"),
-    ] {
-        insert(
-            fields,
-            target,
-            scalar(object.get(source)).unwrap_or_default(),
-        );
-    }
-    Ok(provider_id)
-}
-
-fn normalize_group(
-    object: &Map<String, Value>,
-    fields: &mut BTreeMap<String, String>,
-) -> Result<String, GoogleWorkspaceError> {
-    let provider_id = first_values(object, &["id", "email"])
-        .ok_or(GoogleWorkspaceError::MissingRecordIdentity)?;
-    let email = scalar(object.get("email")).unwrap_or_default();
-    let name = scalar(object.get("name")).unwrap_or_default();
-    insert(
-        fields,
-        "group_id",
-        scalar(object.get("id")).unwrap_or_default(),
-    );
-    insert(fields, "group_email", email.clone());
-    insert(fields, "email", email);
-    insert(fields, "group_name", name.clone());
-    insert(fields, "name", name);
-    for (source, target) in [
-        ("description", "description"),
-        ("adminCreated", "admin_created"),
-        ("directMembersCount", "direct_members_count"),
-    ] {
-        insert(
-            fields,
-            target,
-            scalar(object.get(source)).unwrap_or_default(),
-        );
-    }
-    Ok(provider_id)
-}
-
-fn normalize_group_member(
-    object: &Map<String, Value>,
-    fields: &mut BTreeMap<String, String>,
-    group_key: Option<&str>,
-) -> Result<String, GoogleWorkspaceError> {
-    let group_key = group_key.ok_or(GoogleWorkspaceError::MissingGroupKey)?;
-    let member_id = first_values(object, &["id", "email"])
-        .ok_or(GoogleWorkspaceError::MissingRecordIdentity)?;
-    let email = scalar(object.get("email")).unwrap_or_default();
-    insert(fields, "group_id", group_key.to_owned());
-    insert(fields, "group_email", group_key.to_owned());
-    insert(
-        fields,
-        "member_id",
-        scalar(object.get("id")).unwrap_or_default(),
-    );
-    insert(fields, "member_email", email.clone());
-    insert(
-        fields,
-        "member_user_id",
-        scalar(object.get("id")).unwrap_or_default(),
-    );
-    insert(fields, "email", email);
-    insert(fields, "member_type", lowercase_scalar(object.get("type")));
-    insert(
-        fields,
-        "role",
-        scalar(object.get("role")).unwrap_or_default(),
-    );
-    insert(
-        fields,
-        "member_status",
-        scalar(object.get("status")).unwrap_or_default(),
-    );
-    insert(
-        fields,
-        "user_id",
-        scalar(object.get("id")).unwrap_or_default(),
-    );
-    Ok(format!("{group_key}::{member_id}"))
-}
-
-fn normalize_role_assignment(
-    object: &Map<String, Value>,
-    fields: &mut BTreeMap<String, String>,
-    users: &BTreeMap<String, ResolvedUser>,
-) -> Result<String, GoogleWorkspaceError> {
-    let provider_id = scalar(object.get("roleAssignmentId"))
-        .filter(|value| !value.is_empty())
-        .ok_or(GoogleWorkspaceError::MissingRecordIdentity)?;
-    let assigned_to = scalar(object.get("assignedTo")).unwrap_or_default();
-    let resolved = users.get(&assigned_to);
-    let subject_email = resolved
-        .map(|user| user.email.clone())
-        .filter(|value| !value.is_empty())
-        .or_else(|| email_like(&assigned_to))
-        .unwrap_or_default();
-    insert(fields, "role_assignment_id", provider_id.clone());
-    insert(
-        fields,
-        "role_id",
-        scalar(object.get("roleId")).unwrap_or_default(),
-    );
-    insert(fields, "subject_email", subject_email.clone());
-    insert(fields, "subject_id", assigned_to.clone());
-    insert(fields, "subject_login", subject_email);
-    insert(
-        fields,
-        "subject_name",
-        resolved.map(|user| user.name.clone()).unwrap_or_default(),
-    );
-    insert(fields, "assigned_to", assigned_to);
-    let subject_type = lowercase_scalar(object.get("assigneeType"));
-    insert(fields, "subject_type", subject_type.clone());
-    insert(fields, "principal_type", subject_type);
-    insert(
-        fields,
-        "scope_type",
-        scalar(object.get("scopeType")).unwrap_or_default(),
-    );
-    insert(
-        fields,
-        "org_unit_id",
-        scalar(object.get("orgUnitId")).unwrap_or_default(),
-    );
-    insert(fields, "event_type", "admin.role.assignment".to_owned());
-    insert(fields, "action", "admin.role.assignment".to_owned());
-    Ok(provider_id)
-}
-
-fn normalize_audit(
-    object: &Map<String, Value>,
-    fields: &mut BTreeMap<String, String>,
-) -> Result<String, GoogleWorkspaceError> {
-    let event = object
-        .get("events")
-        .and_then(Value::as_array)
-        .and_then(|events| events.first())
-        .and_then(Value::as_object);
-    let event_name = event
-        .and_then(|event| scalar(event.get("name")))
-        .unwrap_or_default();
-    let event_type = event
-        .and_then(|event| scalar(event.get("type")))
-        .unwrap_or_default();
-    let provider_id = nested_scalar(object, &["id", "uniqueQualifier"])
-        .or_else(|| (!event_name.is_empty()).then(|| event_name.clone()))
-        .or_else(|| nested_scalar(object, &["id", "time"]))
-        .ok_or(GoogleWorkspaceError::MissingRecordIdentity)?;
-    let parameters = audit_parameters(event);
-    let resource_id = first_nonempty([
-        parameters.get("USER_EMAIL").cloned(),
-        parameters.get("GROUP_EMAIL").cloned(),
-        parameters.get("APP_NAME").cloned(),
-        parameters.get("CLIENT_ID").cloned(),
-        parameters.get("ROLE_NAME").cloned(),
-        Some(event_name.clone()),
-    ])
-    .unwrap_or_default();
-    let resource_type = first_nonempty([
-        parameters.get("RESOURCE_TYPE").cloned(),
-        Some(event_type),
-        Some("security_setting".to_owned()),
-    ])
-    .unwrap_or_default();
-    insert(fields, "event_type", event_name.clone());
-    insert(fields, "event_name", event_name.clone());
-    insert(fields, "action", event_name);
-    insert(fields, "resource_id", resource_id.clone());
-    insert(
-        fields,
-        "resource_type",
-        normalize_resource_type(&resource_type),
-    );
-    insert(fields, "resource_name", resource_id);
-    let actor_email = nested_scalar(object, &["actor", "email"]).unwrap_or_default();
-    insert(fields, "actor_email", actor_email.clone());
-    insert(
-        fields,
-        "actor_id",
-        nested_scalar(object, &["actor", "profileId"]).unwrap_or_default(),
-    );
-    insert(fields, "actor_alternate_id", actor_email);
-    insert(
-        fields,
-        "application",
-        nested_scalar(object, &["id", "applicationName"]).unwrap_or_default(),
-    );
-    insert(
-        fields,
-        "customer_id",
-        nested_scalar(object, &["id", "customerId"]).unwrap_or_default(),
-    );
-    Ok(provider_id)
-}
-
-fn audit_parameters(event: Option<&Map<String, Value>>) -> BTreeMap<String, String> {
-    let mut parameters = BTreeMap::new();
-    let Some(values) = event
-        .and_then(|event| event.get("parameters"))
-        .and_then(Value::as_array)
-    else {
-        return parameters;
-    };
-    for parameter in values.iter().filter_map(Value::as_object) {
-        if let (Some(name), Some(value)) = (
-            scalar(parameter.get("name")),
-            scalar(parameter.get("value")),
-        ) {
-            parameters.insert(name.trim().to_uppercase(), value.trim().to_owned());
-        }
-    }
-    parameters
-}
-
-fn first_values(object: &Map<String, Value>, keys: &[&str]) -> Option<String> {
-    keys.iter()
-        .find_map(|key| scalar(object.get(*key)).filter(|value| !value.is_empty()))
-}
-
-fn nested_scalar(object: &Map<String, Value>, path: &[&str]) -> Option<String> {
-    let (first, rest) = path.split_first()?;
-    let mut value = object.get(*first)?;
-    for segment in rest {
-        value = value.as_object()?.get(*segment)?;
-    }
-    scalar(Some(value))
-}
-
-fn scalar(value: Option<&Value>) -> Option<String> {
-    match value? {
-        Value::String(value) => Some(value.trim().to_owned()),
-        Value::Number(value) => Some(value.to_string()),
-        Value::Bool(value) => Some(value.to_string()),
-        Value::Null | Value::Array(_) | Value::Object(_) => None,
-    }
-}
-
-fn lowercase_scalar(value: Option<&Value>) -> String {
-    scalar(value).unwrap_or_default().to_lowercase()
-}
-
-fn insert(fields: &mut BTreeMap<String, String>, key: &str, value: String) {
-    if !value.trim().is_empty() {
-        fields.insert(key.to_owned(), value);
-    }
-}
-
-fn email_like(value: &str) -> Option<String> {
-    let value = value.trim();
-    value.contains('@').then(|| value.to_lowercase())
-}
-
-fn normalize_resource_type(value: &str) -> String {
-    value.trim().to_lowercase().replace([' ', '-'], "_")
-}
-
-fn parse_observed_at(value: &str) -> Result<OffsetDateTime, GoogleWorkspaceError> {
-    OffsetDateTime::parse(value.trim(), &Rfc3339)
-        .map(|value| value.to_offset(UtcOffset::UTC))
-        .map_err(|_| GoogleWorkspaceError::InvalidObservedAt)
-}
-
-fn provider_time(value: Option<String>) -> Option<OffsetDateTime> {
-    let value = value?;
-    let value = value.trim();
-    if value.is_empty() || value == "1970-01-01T00:00:00.000Z" {
-        return None;
-    }
-    OffsetDateTime::parse(value, &Rfc3339)
-        .ok()
-        .map(|value| value.to_offset(UtcOffset::UTC))
-}
-
-fn first_provider_time<const N: usize>(values: [Option<String>; N]) -> Option<OffsetDateTime> {
-    values.into_iter().find_map(provider_time)
-}
-
-fn format_time(value: OffsetDateTime) -> Result<String, GoogleWorkspaceError> {
-    value
-        .to_offset(UtcOffset::UTC)
-        .format(&Rfc3339)
-        .map_err(|_| GoogleWorkspaceError::InvalidObservedAt)
-}
-
-fn materialize_event(
-    record: &GoogleWorkspaceRecord,
-    observed_at: OffsetDateTime,
-) -> Result<GoogleWorkspaceEvent, GoogleWorkspaceError> {
-    let family = GoogleWorkspaceFamily::from_str(&record.family)?;
-    let object = record
-        .payload
-        .as_object()
-        .ok_or(GoogleWorkspaceError::InvalidRecord)?;
-    let occurred_at = match family {
-        GoogleWorkspaceFamily::User => first_provider_time([
-            scalar(object.get("lastLoginTime")),
-            scalar(object.get("creationTime")),
-        ])
-        .unwrap_or(observed_at),
-        GoogleWorkspaceFamily::Audit => {
-            provider_time(nested_scalar(object, &["id", "time"])).unwrap_or(observed_at)
-        }
-        GoogleWorkspaceFamily::Group
-        | GoogleWorkspaceFamily::GroupMember
-        | GoogleWorkspaceFamily::RoleAssignment => observed_at,
-    };
-    let event_id = go_event_id(family, object, &record.fields, occurred_at)?;
-    let discovery_id = first_nonempty([
-        record.fields.get("user_id").cloned(),
-        record.fields.get("group_id").cloned(),
-        record.fields.get("role_assignment_id").cloned(),
-        record.fields.get("event_type").cloned(),
-    ])
-    .ok_or(GoogleWorkspaceError::MissingDiscoveryIdentity)?;
-    let domain = record
-        .fields
-        .get("domain")
-        .cloned()
-        .filter(|value| !value.trim().is_empty())
-        .ok_or(GoogleWorkspaceError::MissingDomain)?;
-    let mut payload = record.payload.clone();
-    let payload_object = payload
-        .as_object_mut()
-        .ok_or(GoogleWorkspaceError::InvalidRecord)?;
-    payload_object.insert("domain".to_owned(), Value::String(domain.clone()));
-    if family == GoogleWorkspaceFamily::GroupMember {
-        let group_key = record
-            .fields
-            .get("group_id")
-            .cloned()
-            .filter(|value| !value.trim().is_empty())
-            .ok_or(GoogleWorkspaceError::MissingGroupKey)?;
-        payload_object.insert("group_key".to_owned(), Value::String(group_key));
-    }
-    Ok(GoogleWorkspaceEvent {
-        event_id,
-        tenant_id: domain.clone(),
-        source_id: SOURCE_ID.to_owned(),
-        provider_kind: family.provider_kind().to_owned(),
-        schema_ref: family.schema_ref().to_owned(),
-        occurred_at: format_time(occurred_at)?,
-        discovery_urn: format!(
-            "urn:cerebro:{domain}:google_workspace_{}:{discovery_id}",
-            family.as_str()
-        ),
-        attributes: record.fields.clone(),
-        payload,
-    })
-}
-
-fn go_event_id(
-    family: GoogleWorkspaceFamily,
-    object: &Map<String, Value>,
-    fields: &BTreeMap<String, String>,
-    occurred_at: OffsetDateTime,
-) -> Result<String, GoogleWorkspaceError> {
-    let (prefix, identity) = match family {
-        GoogleWorkspaceFamily::User => (
-            "google-workspace-user-",
-            first_values(object, &["id", "primaryEmail"]),
-        ),
-        GoogleWorkspaceFamily::Group => (
-            "google-workspace-group-",
-            first_values(object, &["id", "email"]),
-        ),
-        GoogleWorkspaceFamily::GroupMember => {
-            let group_key = fields.get("group_id").cloned();
-            let member_id = first_values(object, &["id", "email"]);
-            (
-                "google-workspace-group-member-",
-                group_key
-                    .zip(member_id)
-                    .map(|(group, member)| format!("{group}-{member}")),
-            )
-        }
-        GoogleWorkspaceFamily::RoleAssignment => (
-            "google-workspace-role-assignment-",
-            scalar(object.get("roleAssignmentId")),
-        ),
-        GoogleWorkspaceFamily::Audit => (
-            "google-workspace-audit-",
-            first_nonempty([
-                nested_scalar(object, &["id", "uniqueQualifier"]),
-                fields.get("event_type").cloned(),
-                Some((occurred_at.unix_timestamp_nanos() / 1_000_000).to_string()),
-            ]),
-        ),
-    };
-    identity
-        .filter(|value| !value.trim().is_empty())
-        .map(|identity| format!("{prefix}{identity}"))
-        .ok_or(GoogleWorkspaceError::MissingRecordIdentity)
-}
-
-fn first_nonempty<const N: usize>(values: [Option<String>; N]) -> Option<String> {
-    values
-        .into_iter()
-        .flatten()
-        .find(|value| !value.trim().is_empty())
-        .map(|value| value.trim().to_owned())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    const AUDIT_FIXTURE: &[u8] = include_bytes!(concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../../sources/googleworkspace/testdata/read_audit.json"
-    ));
-    const GROUP_FIXTURE: &[u8] = include_bytes!(concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../../sources/googleworkspace/testdata/read_group.json"
-    ));
-    const GROUP_MEMBER_FIXTURE: &[u8] = include_bytes!(concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../../sources/googleworkspace/testdata/read_group_member.json"
-    ));
-    const ROLE_FIXTURE: &[u8] = include_bytes!(concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../../sources/googleworkspace/testdata/read_role_assignment.json"
-    ));
-    const USER_FIXTURE: &[u8] = include_bytes!(concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/../../sources/googleworkspace/testdata/read_user.json"
-    ));
-
-    fn kernel(family: GoogleWorkspaceFamily) -> GoogleWorkspaceKernel {
-        GoogleWorkspaceKernel::new(
-            "https://admin.googleapis.com",
-            "writer.com",
-            family,
-            GoogleWorkspaceFilters {
-                customer_id: Some("C01".to_owned()),
-                group_key: Some("security@writer.com".to_owned()),
-                application: Some("admin".to_owned()),
-            },
-            Some(2),
-        )
-        .unwrap()
-    }
-
-    fn fixture_page(family: GoogleWorkspaceFamily, fixture: &[u8]) -> Vec<u8> {
-        let records: Value = serde_json::from_slice(fixture).unwrap();
-        serde_json::to_vec(&serde_json::json!({family.response_field(): records})).unwrap()
-    }
-
-    fn decoded_fixture_page(family: GoogleWorkspaceFamily, fixture: &[u8]) -> GoogleWorkspacePage {
-        let kernel = kernel(family);
-        let request = kernel.plan(None).unwrap();
-        let outcome = kernel
-            .decode(&request, &fixture_page(family, fixture))
-            .unwrap();
-        let outcome = match outcome {
-            GoogleWorkspaceOutcome::Request(request) => kernel
-                .decode(
-                    &request,
-                    br#"{"id":"1001","primaryEmail":"admin@writer.com","name":{"fullName":"Admin Writer"}}"#,
-                )
-                .unwrap(),
-            page => page,
-        };
-        let GoogleWorkspaceOutcome::Page(page) = outcome else {
-            panic!("fixture required more than one bounded lookup")
-        };
-        page
-    }
-
-    #[test]
-    fn all_five_families_plan_exact_paths_queries_and_auth_contract() {
-        let cases = [
-            (
-                GoogleWorkspaceFamily::Audit,
-                "/admin/reports/v1/activity/users/all/applications/admin",
-                Some(("customerId", "C01")),
-            ),
-            (
-                GoogleWorkspaceFamily::Group,
-                "/admin/directory/v1/groups",
-                Some(("customer", "C01")),
-            ),
-            (
-                GoogleWorkspaceFamily::GroupMember,
-                "/admin/directory/v1/groups/security@writer.com/members",
-                None,
-            ),
-            (
-                GoogleWorkspaceFamily::RoleAssignment,
-                "/admin/directory/v1/customer/C01/roleassignments",
-                None,
-            ),
-            (
-                GoogleWorkspaceFamily::User,
-                "/admin/directory/v1/users",
-                Some(("customer", "C01")),
-            ),
-        ];
-        for (family, path, scope) in cases {
-            let request = kernel(family).plan(Some("page-2")).unwrap();
-            let query = request
-                .url()
-                .query_pairs()
-                .into_owned()
-                .collect::<BTreeMap<_, _>>();
-            assert_eq!(request.url().path(), path);
-            assert_eq!(query.get("maxResults").map(String::as_str), Some("2"));
-            assert_eq!(query.get("pageToken").map(String::as_str), Some("page-2"));
-            if let Some((key, value)) = scope {
-                assert_eq!(query.get(key).map(String::as_str), Some(value));
-            }
-            assert_eq!(request.authorization_scheme(), "Bearer");
-            assert_eq!(request.accept(), "application/json");
-        }
-    }
-
-    #[test]
-    fn go_fixtures_decode_for_all_families_with_source_specific_identities() {
-        let cases = [
-            (GoogleWorkspaceFamily::Audit, AUDIT_FIXTURE, "audit-1", 2),
-            (GoogleWorkspaceFamily::Group, GROUP_FIXTURE, "group-1", 1),
-            (
-                GoogleWorkspaceFamily::GroupMember,
-                GROUP_MEMBER_FIXTURE,
-                "security@writer.com::member-1",
-                2,
-            ),
-            (GoogleWorkspaceFamily::User, USER_FIXTURE, "1001", 2),
-        ];
-        for (family, fixture, provider_id, count) in cases {
-            let kernel = kernel(family);
-            let request = kernel.plan(None).unwrap();
-            let outcome = kernel
-                .decode(&request, &fixture_page(family, fixture))
-                .unwrap();
-            let GoogleWorkspaceOutcome::Page(page) = outcome else {
-                panic!("expected direct fixture page")
-            };
-            assert_eq!(page.records.len(), count);
-            assert_eq!(page.records[0].provider_id, provider_id);
-            assert_eq!(page.records[0].provider_kind, family.provider_kind());
-            assert_eq!(
-                page.records[0].fields.get("domain").map(String::as_str),
-                Some("writer.com")
-            );
-        }
-    }
-
-    #[test]
-    fn provider_cursor_and_user_attributes_preserve_go_contract() {
-        let kernel = kernel(GoogleWorkspaceFamily::User);
-        let request = kernel.plan(None).unwrap();
-        let records: Value = serde_json::from_slice(USER_FIXTURE).unwrap();
-        let body = serde_json::to_vec(&serde_json::json!({
-            "users": records,
-            "nextPageToken": "page-2"
-        }))
-        .unwrap();
-        let GoogleWorkspaceOutcome::Page(page) = kernel.decode(&request, &body).unwrap() else {
-            panic!("expected user page")
-        };
-        assert_eq!(page.next_cursor.as_deref(), Some("page-2"));
-        assert_eq!(
-            page.records[0].fields.get("email").map(String::as_str),
-            Some("admin@writer.com")
-        );
-        assert_eq!(
-            page.records[0]
-                .fields
-                .get("mfa_enrolled")
-                .map(String::as_str),
-            Some("false")
-        );
-        let materialized = page.materialize("2026-08-20T20:00:00-07:00").unwrap();
-        assert_eq!(materialized.next_cursor.as_deref(), Some("page-2"));
-        assert_eq!(materialized.checkpoint_cursor.as_deref(), Some("page-2"));
-        assert_eq!(
-            materialized.watermark.as_deref(),
-            Some("2026-04-23T00:00:00Z")
-        );
-    }
-
-    #[test]
-    fn all_five_families_materialize_go_event_and_discovery_contracts() {
-        let cases = [
-            (
-                GoogleWorkspaceFamily::Audit,
-                AUDIT_FIXTURE,
-                "google-workspace-audit-audit-1",
-                "google_workspace/audit/v1",
-                "2026-04-23T00:00:00Z",
-                "urn:cerebro:writer.com:google_workspace_audit:CHANGE_TWO_STEP_VERIFICATION_ENFORCEMENT",
-                "google-workspace-audit-audit-2",
-            ),
-            (
-                GoogleWorkspaceFamily::Group,
-                GROUP_FIXTURE,
-                "google-workspace-group-group-1",
-                "google_workspace/group/v1",
-                "2026-08-21T03:00:00Z",
-                "urn:cerebro:writer.com:google_workspace_group:group-1",
-                "google-workspace-group-group-1",
-            ),
-            (
-                GoogleWorkspaceFamily::GroupMember,
-                GROUP_MEMBER_FIXTURE,
-                "google-workspace-group-member-security@writer.com-member-1",
-                "google_workspace/group_member/v1",
-                "2026-08-21T03:00:00Z",
-                "urn:cerebro:writer.com:google_workspace_group_member:member-1",
-                "google-workspace-group-member-security@writer.com-member-2",
-            ),
-            (
-                GoogleWorkspaceFamily::RoleAssignment,
-                ROLE_FIXTURE,
-                "google-workspace-role-assignment-ra-1",
-                "google_workspace/role_assignment/v1",
-                "2026-08-21T03:00:00Z",
-                "urn:cerebro:writer.com:google_workspace_role_assignment:ra-1",
-                "google-workspace-role-assignment-ra-1",
-            ),
-            (
-                GoogleWorkspaceFamily::User,
-                USER_FIXTURE,
-                "google-workspace-user-1001",
-                "google_workspace/user/v1",
-                "2025-01-15T00:00:00Z",
-                "urn:cerebro:writer.com:google_workspace_user:1001",
-                "google-workspace-user-1002",
-            ),
-        ];
-        for (family, fixture, event_id, schema_ref, occurred_at, discovery_urn, checkpoint) in cases
-        {
-            let page = decoded_fixture_page(family, fixture)
-                .materialize("2026-08-20T20:00:00-07:00")
-                .unwrap();
-            assert_eq!(page.events[0].event_id, event_id);
-            assert_eq!(page.events[0].source_id, "google_workspace");
-            assert_eq!(page.events[0].tenant_id, "writer.com");
-            assert_eq!(page.events[0].provider_kind, family.provider_kind());
-            assert_eq!(page.events[0].schema_ref, schema_ref);
-            assert_eq!(page.events[0].occurred_at, occurred_at);
-            assert_eq!(page.events[0].discovery_urn, discovery_urn);
-            assert_eq!(
-                page.events[0].payload.get("domain").and_then(Value::as_str),
-                Some("writer.com")
-            );
-            assert_eq!(page.checkpoint_cursor.as_deref(), Some(checkpoint));
-            assert_eq!(
-                page.watermark,
-                page.events.last().map(|event| event.occurred_at.clone())
-            );
-        }
-        let memberships =
-            decoded_fixture_page(GoogleWorkspaceFamily::GroupMember, GROUP_MEMBER_FIXTURE)
-                .materialize("2026-08-20T20:00:00-07:00")
-                .unwrap();
-        assert_eq!(
-            memberships.events[0]
-                .payload
-                .get("group_key")
-                .and_then(Value::as_str),
-            Some("security@writer.com")
-        );
-    }
-
-    #[test]
-    fn materialization_matches_go_time_fallback_and_empty_page_semantics() {
-        let kernel = kernel(GoogleWorkspaceFamily::User);
-        let request = kernel.plan(None).unwrap();
-        let GoogleWorkspaceOutcome::Page(page) = kernel
-            .decode(
-                &request,
-                br#"{"users":[{"id":"1003","primaryEmail":"fallback@writer.com","lastLoginTime":"invalid","creationTime":"2026-08-20T01:02:03.456Z"}]}"#,
-            )
-            .unwrap()
-        else {
-            panic!("expected user page")
-        };
-        let materialized = page.materialize("2026-08-21T03:00:00Z").unwrap();
-        assert_eq!(
-            materialized.events[0].occurred_at,
-            "2026-08-20T01:02:03.456Z"
-        );
-        assert!(matches!(
-            page.materialize("not-a-time"),
-            Err(GoogleWorkspaceError::InvalidObservedAt)
-        ));
-
-        let empty = GoogleWorkspacePage {
-            records: Vec::new(),
-            next_cursor: Some("provider-should-not-advance".to_owned()),
-        }
-        .materialize("2026-08-21T03:00:00Z")
-        .unwrap();
-        assert!(empty.events.is_empty());
-        assert_eq!(empty.next_cursor, None);
-        assert_eq!(empty.checkpoint_cursor, None);
-        assert_eq!(empty.watermark, None);
-    }
-
-    #[test]
-    fn materialization_fail_closes_go_read_only_identity_fallbacks() {
-        for (family, body) in [
-            (
-                GoogleWorkspaceFamily::User,
-                br#"{"users":[{"primaryEmail":"fallback@writer.com"}]}"#.as_slice(),
-            ),
-            (
-                GoogleWorkspaceFamily::Group,
-                br#"{"groups":[{"email":"fallback-group@writer.com"}]}"#.as_slice(),
-            ),
-            (
-                GoogleWorkspaceFamily::Audit,
-                br#"{"items":[{"id":{"time":"2026-08-20T01:02:03Z","uniqueQualifier":"audit-without-event"},"events":[{"name":""}]}]}"#.as_slice(),
-            ),
-        ] {
-            let kernel = kernel(family);
-            let request = kernel.plan(None).unwrap();
-            let GoogleWorkspaceOutcome::Page(page) = kernel.decode(&request, body).unwrap()
-            else {
-                panic!("expected direct page")
-            };
-            assert!(matches!(
-                page.materialize("2026-08-21T03:00:00Z"),
-                Err(GoogleWorkspaceError::MissingDiscoveryIdentity)
-            ));
-        }
-    }
-
-    #[test]
-    fn missing_role_assignment_id_and_malformed_records_fail_before_admission() {
-        let role_kernel = kernel(GoogleWorkspaceFamily::RoleAssignment);
-        let request = role_kernel.plan(None).unwrap();
-        assert!(matches!(
-            role_kernel.decode(
-                &request,
-                br#"{"items":[{"roleId":"super-admin","assigneeType":"GROUP"}]}"#,
-            ),
-            Err(GoogleWorkspaceError::MissingRecordIdentity)
-        ));
-
-        let user_kernel = kernel(GoogleWorkspaceFamily::User);
-        let request = user_kernel.plan(None).unwrap();
-        assert!(matches!(
-            user_kernel.decode(&request, br#"{"users":[42]}"#),
-            Err(GoogleWorkspaceError::InvalidRecord)
-        ));
-    }
-
-    #[test]
-    fn role_assignment_fanout_caches_unique_user_once_per_page() {
-        let kernel = kernel(GoogleWorkspaceFamily::RoleAssignment);
-        let request = kernel.plan(None).unwrap();
-        let mut records: Vec<Value> = serde_json::from_slice(ROLE_FIXTURE).unwrap();
-        let mut second = records[0].clone();
-        second["roleAssignmentId"] = Value::String("ra-2".to_owned());
-        records.push(second);
-        let body = serde_json::to_vec(&serde_json::json!({
-            "items": records,
-            "nextPageToken": "roles-2"
-        }))
-        .unwrap();
-        let GoogleWorkspaceOutcome::Request(user_request) = kernel.decode(&request, &body).unwrap()
-        else {
-            panic!("expected role user lookup")
-        };
-        assert_eq!(user_request.url().path(), "/admin/directory/v1/users/1001");
-        let GoogleWorkspaceOutcome::Page(page) = kernel
-            .decode(
-                &user_request,
-                br#"{"id":"1001","primaryEmail":"admin@writer.com","name":{"fullName":"Admin Writer"}}"#,
-            )
-            .unwrap()
-        else {
-            panic!("expected enriched role page")
-        };
-        assert_eq!(page.records.len(), 2);
-        assert_eq!(page.next_cursor.as_deref(), Some("roles-2"));
-        assert!(page.records.iter().all(|record| {
-            record.fields.get("subject_email").map(String::as_str) == Some("admin@writer.com")
-        }));
-        assert!(page.records.iter().all(|record| {
-            record.fields.get("subject_name").map(String::as_str) == Some("Admin Writer")
-        }));
-    }
-
-    #[test]
-    fn genuine_role_fixture_uses_one_bounded_lookup() {
-        let kernel = kernel(GoogleWorkspaceFamily::RoleAssignment);
-        let request = kernel.plan(None).unwrap();
-        let outcome = kernel
-            .decode(
-                &request,
-                &fixture_page(GoogleWorkspaceFamily::RoleAssignment, ROLE_FIXTURE),
-            )
-            .unwrap();
-        let GoogleWorkspaceOutcome::Request(request) = outcome else {
-            panic!("expected fixture lookup request")
-        };
-        assert_eq!(request.url().path(), "/admin/directory/v1/users/1001");
-    }
-
-    #[test]
-    fn kernel_fails_closed_on_unsafe_origins_invalid_scope_and_bad_pages() {
-        assert!(matches!(
-            GoogleWorkspaceKernel::new(
-                "http://169.254.169.254",
-                "writer.com",
-                GoogleWorkspaceFamily::User,
-                GoogleWorkspaceFilters::default(),
-                None,
-            ),
-            Err(GoogleWorkspaceError::InvalidBaseUrl)
-        ));
-        assert!(matches!(
-            GoogleWorkspaceKernel::new(
-                "https://admin.googleapis.com",
-                "writer.com",
-                GoogleWorkspaceFamily::GroupMember,
-                GoogleWorkspaceFilters::default(),
-                None,
-            ),
-            Err(GoogleWorkspaceError::MissingGroupKey)
-        ));
-        let user_kernel = kernel(GoogleWorkspaceFamily::User);
-        let group_request = kernel(GoogleWorkspaceFamily::Group).plan(None).unwrap();
-        assert!(matches!(
-            user_kernel.decode(&group_request, br#"{"groups":[]}"#),
-            Err(GoogleWorkspaceError::RequestScopeMismatch)
-        ));
-        let request = user_kernel.plan(None).unwrap();
-        assert!(matches!(
-            user_kernel.decode(&request, br#"{"users":{}}"#),
-            Err(GoogleWorkspaceError::InvalidResponse)
-        ));
-    }
-}
